+++ src/minify/css/MinifyCssCommand.ts
@@ -21,12 +21,16 @@ export class MinifyCssCommand
     public execute()
     {
         StatusBarMessage.hideError();
-        let opts = {
+        let opts:any = {
             processImport: false,
             rebase: false,
             advanced: true,
-            groupmedia: false
+            groupmedia: false,
+            sourceMap: false,
+            sourceMapFileInline: false
         }
+        const cssFile = this.document.fileName;
+        const cssPath: string = path.dirname(cssFile);
 
         let globalOptions = Configuration.getGlobalOptions(this.document.fileName, 'css');
         let compilingMessage = StatusBarMessage.show("$(zap) Minifing css", StatusBarMessageTypes.INDEFINITE);
@@ -35,13 +39,55 @@ export class MinifyCssCommand
 
         readFilePromise(this.document.fileName).then(buffer =>
             {
+                const outExt = opts.outExt?opts.outExt:'.css';
+                const baseFilename: string = path.parse(cssFile).name;
+
+                let outDir = cssPath;
+                if(opts.outDir){
+                    outDir = resolveFilePath(opts.outDir, cssPath, cssFile);
+                }
+                const outFile = path.resolve(outDir, baseFilename+outExt);
+
                 let content: string = buffer.toString();
                 if(opts.groupmedia){
                     let grouper = require('group-css-media-queries');
                     content = grouper(content);
                 }
                 let output = new CleanCSS(opts).minify(content);
-                return writeFileContents(this.document.fileName, output.styles);
+                let css = output.styles;
+
+                let sourceMapFile: string | undefined, sourceMap;
+                let replaceList:any = {
+                    "$stdin": path.relative(outDir, cssFile)
+                };
+                if (opts.sourceMap && output.sourceMap)
+                {
+                    if(!opts.sourceMapFileInline){
+                        sourceMapFile = outFile + '.map';
+                    }
+
+                    sourceMap = JSON.parse(output.sourceMap);
+                    let x;
+                    for(x in sourceMap.sources){
+                        let path = sourceMap.sources[x];
+                        if(replaceList[path]) sourceMap.sources[x] = replaceList[path];
+                    }
+
+                    if(sourceMapFile){
+                        const mapFileUrl: string = path.basename(sourceMapFile);
+                        css += "\n"+'/*# sourceMappingURL='+mapFileUrl+' */';
+                    }else{
+                        css += "\n"+'/*# sourceMappingURL=data:application/json;charset=utf-8;base64,'+Buffer.from(JSON.stringify(sourceMap)).toString("base64")+'  */';
+                    }
+                }
+
+                return writeFileContents(outFile, css).then(() =>
+                {
+                    if (sourceMapFile && sourceMap)
+                    {
+                        return writeFileContents(sourceMapFile, JSON.stringify(sourceMap));
+                    }
+                });
             }).then(() =>
             {
                 compilingMessage.dispose();
@@ -98,4 +144,24 @@ function readFilePromise(this: void, filename: string): Promise<Buffer>
             }
         });
     });
+}
+
+function intepolatePath(this: void, path: string): string
+{
+    if(vscode.workspace.workspaceFolders){
+        let rootPath = vscode.workspace.workspaceFolders[0];
+        return (<string>path).replace(/\$\{workspaceRoot\}/g, rootPath.uri.path);
+    }
+    return path;
+}
+
+function resolveFilePath(this: void, main: string, tsPath: string, currentTsFile: string): string
+{
+    const interpolatedFilePath: string = intepolatePath(main);
+    const resolvedFilePath: string = path.resolve(tsPath, interpolatedFilePath);
+    if (resolvedFilePath.indexOf(currentTsFile) >= 0)
+    {
+        return ''; // avoid infinite loops
+    }
+    return resolvedFilePath;
 }